--- University-Web-App-Development/src/main/java/com/university/Entity/Student.java
@@ -1,58 +1,0 @@
-package com.university.Entity;
-
-import jakarta.persistence.Entity;
-
-import jakarta.persistence.Id;
-
-
-@Entity
-public class Student {
-
-    @Id
-    private String studentId;
-    private String country;
-    private String university;
-    private String name;
-    private String email;
-
-
-    public String getStudentId() {
-        return studentId;
-    }
-
-    public void setStudentId(String studentId) {
-        this.studentId = studentId;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    public String getUniversity() {
-        return university;
-    }
-
-    public void setUniversity(String university) {
-        this.university = university;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
