+++ University-Web-App-Development/src/main/java/com/university/controller/LoginController.java
@@ -1,7 +1,7 @@
 package com.university.controller;
 
 import com.university.DTO.LoginResponse;
-import com.university.Entity.Student;
+import com.university.model.Admin;
 import com.university.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,26 +18,46 @@ public class LoginController {
     private StudentRepository studentRepository;
 
     @PostMapping("/login")
-    public ResponseEntity<LoginResponse> login(@RequestBody Student student) {
-
-
+    public ResponseEntity<LoginResponse> login(@RequestBody Admin.Student student) {
         System.out.println("Login attempt for student: " + student.getEmail());
 
-
-        Student existingStudent = studentRepository.findByEmail(student.getEmail());
+        Admin.Student existingStudent = studentRepository.findByEmail(student.getEmail());
 
         if (existingStudent != null) {
-
             return ResponseEntity.status(HttpStatus.OK)
                     .body(new LoginResponse("success", existingStudent.getStudentId(), "success"));
         }
 
-
         student.setStudentId(UUID.randomUUID().toString());
         studentRepository.save(student);
 
-
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(new LoginResponse("success", student.getStudentId(), "success"));
+    }
+
+
+    @GetMapping("/{studentId}")
+    public ResponseEntity<Admin.Student> getStudent(@PathVariable String studentId) {
+        System.out.println("Fetching student details for studentId: " + studentId);
+
+        Admin.Student student = studentRepository.findByStudentId(studentId);
+        if (student != null) {
+            return ResponseEntity.status(HttpStatus.OK).body(student);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+    }
+
+
+    @GetMapping("/email/{email}")
+    public ResponseEntity<Admin.Student> getStudentByEmail(@PathVariable String email) {
+        System.out.println("Fetching student details for email: " + email);
+
+        Admin.Student student = studentRepository.findByEmail(email);
+        if (student != null) {
+            return ResponseEntity.status(HttpStatus.OK).body(student);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
 }
