+++ University-Web-App-Development/src/main/java/com/university/model/Admin.java
@@ -36,4 +36,56 @@ public class Admin {
     public void setPassword(String password) {
         this.password = password;
     }
+
+    @Entity
+    public static class Student {
+
+        @Id
+        private String studentId;
+        private String country;
+        private String university;
+        private String name;
+        private String email;
+
+
+        public String getStudentId() {
+            return studentId;
+        }
+
+        public void setStudentId(String studentId) {
+            this.studentId = studentId;
+        }
+
+        public String getCountry() {
+            return country;
+        }
+
+        public void setCountry(String country) {
+            this.country = country;
+        }
+
+        public String getUniversity() {
+            return university;
+        }
+
+        public void setUniversity(String university) {
+            this.university = university;
+        }
+
+        public String getName() {
+            return name;
+        }
+
+        public void setName(String name) {
+            this.name = name;
+        }
+
+        public String getEmail() {
+            return email;
+        }
+
+        public void setEmail(String email) {
+            this.email = email;
+        }
+    }
 }
